plugins/invoice/lib/requirements.txt
pymupdf>=1.23.0
Pillow>=10.0.0
img2pdf>=0.5.0
pydantic>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
json-repair>=0.25.0
google-genai
openai
anthropic
